feat(networking): decompose CniStatus into NetworkOverview with islands
- use-poll.ts: shared polling hook with error backoff and visibility check
- cilium-types.ts: discriminated union types for Cilium API responses
- CniOverview.tsx: extracted Overview + Health cards with 120s polling
- CiliumConfigEditor.tsx: extracted Configuration tab (Cilium-only)
- NetworkOverview.tsx: layout shell with Status/Config tabs + Cilium grid
- Remove CniStatus.tsx, replace with NetworkOverview in dashboard
- Rename "CNI Plugin" tab to "Overview" in SubNav
